+++ src/sensorthings/admin.py
@@ -1,6 +1,7 @@
 from django.contrib import admin
 from django.apps import apps
 from django.core.exceptions import ImproperlyConfigured
+from django.conf import settings
 
 app_models = apps.get_app_config("sensorthings").get_models()
 
@@ -19,30 +20,8 @@ def load_admin():
         except ImproperlyConfigured:
             pass
 
-if not hasattr(settings, "SENSORTHINGS_DISABLE_MODELS") and not hasattr(
-    settings, "SENSORTHINGS_DISABLE_ADMIN"
-):
+# load admin if not disabled
+if (not hasattr(settings, "SENSORTHINGS_DISABLE_MODELS") or settings.SENSORTHINGS_DISABLE_MODELS is False) and (not hasattr(settings, "SENSORTHINGS_DISABLE_ADMIN") or settings.SENSORTHINGS_DISABLE_ADMIN is False):
     load_admin()
 
-elif (
-    hasattr(settings, "SENSORTHINGS_DISABLE_MODELS")
-    and hasattr(settings, "SENSORTHINGS_DISABLE_ADMIN")
-    and settings.SENSORTHINGS_DISABLE_MODELS is False
-    and settings.SENSORTHINGS_DISABLE_ADMIN is False
-):
-    load_admin()
-
-elif (
-    not hasattr(settings, "SENSORTHINGS_DISABLE_MODELS")
-    and hasattr(settings, "SENSORTHINGS_DISABLE_ADMIN")
-    and settings.SENSORTHINGS_DISABLE_ADMIN is False
-):
-    load_admin()
-
-elif (
-    hasattr(settings, "SENSORTHINGS_DISABLE_MODELS")
-    and settings.SENSORTHINGS_DISABLE_MODELS is False
-    and not hasattr(settings, "SENSORTHINGS_DISABLE_ADMIN")
-):
-    load_admin()